ref(example): move code in main for pub preview
